+++ src/pages/users/ContributionStatsPage/index.tsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import { useParams } from "react-router-dom";
 import { logEvent } from "lib/events";
 import useContributions from "hooks/apiHooks/useContributions";
-import Banner from "components/moleculars/cards/Banner";
+import Banner from "components/moleculars/banners/Banner";
 import YoutubeEmbed from "components/moleculars/YoutubeEmbed";
 import Spinner from "components/atomics/Spinner";
 import { useTranslation } from "react-i18next";
